fully implemented syncing gold

+++ src/main/java/jsonObjects/GoldSync.java
@@ -1,14 +1,14 @@
 package jsonObjects;
 
 public class GoldSync {
-	public float gold;
-	public float goldPerSec;
+	public double gold;
+	public double goldPerSec;
 	
 	public GoldSync() {
 		
 	}
 
-	public GoldSync(float gold, float goldPerSec) {
+	public GoldSync(double gold, double goldPerSec) {
 		super();
 		this.gold = gold;
 		this.goldPerSec = goldPerSec;
